requirements.txt
streamlit
numpy
scikit-learn
google-generativeai
python-dotenv
pandas